requirements.txt
absl-py==0.7.1
anaconda==0.0.1.1
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.1.8
django-bootstrap-dynamic-formsets==0.5.0
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-category==2.0.1
django-extra-views==0.12.0
django-heroku==0.3.1
django-modelcluster==4.4
django-nested-admin==3.2.2
django-object-tools==2.0.0
django-taggit==0.23.0
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
draftjs-exporter==2.1.5
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
html5lib==1.0.1
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.3
olefile==0.46
opencv-python==4.1.0.25
parso==0.4.0
pbr==5.1.3
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
protobuf==3.6.0
psycopg2==2.8.2
Pygments==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-monkey-business==1.0.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
scipy==1.2.1
SimpleITK==1.2.0
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.2
wagtail==2.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
whitenoise==4.1.2
Willow==1.1
